improvement(middleware): remove old middleware since wasn't being used and refactored code
