+++ src/app/api/(routes)/llm/query/route.ts
@@ -1,6 +1,6 @@
-"use server";
-import { NextRequest } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import { makeIterator, iteratorToStream } from "../iterator";
+import { getAuthenticatedUser } from "@/lib/auth/get-authenticated-user";
 
 const getCorsHeaders = (origin: string) => ({
   "Access-Control-Allow-Origin": origin,
@@ -9,18 +9,21 @@ const getCorsHeaders = (origin: string) => ({
   "Access-Control-Allow-Credentials": "true",
 });
 
-export async function OPTIONS(req: NextRequest) {
-  const origin = req.headers.get("origin") || "http://localhost:3000";
-  const corsHeaders = getCorsHeaders(origin);
-  return new Response(null, { status: 204, headers: corsHeaders });
-}
-
 export async function POST(req: NextRequest) {
-  const origin = req.headers.get("origin") || "http://localhost:3000";
+  const origin = req.headers.get("origin") || (process.env.NODE_ENV === 'development' ? '*' : 'https://openjustice.ai');
   const corsHeaders = getCorsHeaders(origin);
+
+  // Check authentication first
+  const decodedToken = await getAuthenticatedUser();
+  if (decodedToken instanceof NextResponse) {
+    const response = decodedToken;
+    Object.entries(corsHeaders).forEach(([key, value]) => {
+      response.headers.set(key, value);
+    });
+    return response;
+  }
 
   const iterator = makeIterator(await req.json());
   const stream = await iteratorToStream(iterator);
-
   return new Response(stream, { headers: corsHeaders });
 }
